tools/transcribe_api/requirements.txt
flask 
python-dotenv 
pydub 
flask_cors
flask 
python-dotenv 
pydub 
flask_cors
